rename and separate plugins to own files
